requirements.txt
Flask==0.12.3 
gunicorn==19.6.0	
numpy==1.14.5
pandas==0.24.0
requests==2.21.0
path.py==11.5.0
torchvision==0.2.1
torch==0.4.1
pillow==5.4.1
